Added AJAX to Safety Check of places
