+++ Visitor/views.py
@@ -1,11 +1,8 @@
-from tkinter import Place
 from django.contrib.auth import authenticate,login, logout
 from django.shortcuts import render, redirect
 from Visitor.models import VisitorDetails,User
 from django.http import HttpResponse, HttpResponseRedirect
-from home import urls
-from Places.models import PlacesDetails, ComplaintsData
-import pickle
+from Places.models import PlacesDetails
 from django.conf import settings
 from twilio.rest import Client
 from django import forms
@@ -119,88 +116,6 @@ def profile(request):
             # 'complaints': u_complaints,
         }
         return render(request, 'home/profile.html', context=data)
-
-
-def safety_check(request):
-        d = request.user
-        all_places = PlacesDetails.objects.all()
-        data = {
-            'places': all_places,
-            'user': d,
-            'status': '1',
-            'this_place': '',
-            'result': '-1',
-            'A': {1, 2, 3, 4, 5, 6},
-        }
-        return render(request,'home/safety_check.html',context=data)
-
-def measure_safety(request,place_id):
-        place= PlacesDetails.objects.get(pk=place_id)
-        #gender= VisitorDetails.objects.get(pk=user_id).gender
-        #place=PlacesDetails.objects.get(name=place_name)
-        d = request.user
-        v= VisitorDetails.objects.get(user=d)
-        gender = v.gender
-        # print(gender)
-        if gender == 'MALE': gender = 0
-        elif gender =='FEMALE': gender = 1
-        else: gender = 2
-        income=place.avg_income
-        density=place.population_density
-        age=place.avg_age
-        policestationcount=place.police_station_count
-        petrolingvans=place.petroling_vans
-        moralitylevel=place.morality_level
-
-        # result = getPrediction(0, 18960,50,50000,10,50, 10)
-        result = getPrediction(gender, density, age, income , policestationcount, petrolingvans, moralitylevel)
-        if result < 0: result=0
-        print("prediction: ",result[0])
-        result[0]*=10
-        result[0]=100-result[0]
-        data = {
-            'place': place,
-            'user': d,
-            'status': '1',
-            'result': round(result[0]),
-            'A': {1, 2, 3, 4, 5, 6},
-        }
-        return render(request,'home/measure_safety.html',context=data)
-
-def getPrediction(gender, density, age, income , policestationcount, petrolingvans, moralitylevel):
-    userdata = [[gender, density, age, income , policestationcount, petrolingvans, moralitylevel]]
-    import os
-    module_dir = os.path.dirname(__file__)  # get current directory
-    file_path = os.path.join(module_dir, 'model2.sav')
-    file2 = open(file_path, 'rb')   
-    m2 = pickle.load(file2)
-    file2.close()
-    prediction2 = m2.predict(userdata)
-    print(prediction2)
-
-    file_path = os.path.join(module_dir, 'model3.sav')
-    file3 = open(file_path, 'rb')   
-    m3 = pickle.load(file3)
-    file3.close()
-    prediction3 = m3.predict(userdata)
-    print(prediction3)
-
-    file_path = os.path.join(module_dir, 'model4.sav')
-    file4 = open(file_path, 'rb')   
-    m4 = pickle.load(file4)
-    file4.close()
-    prediction4 = m4.predict(userdata)
-    print(prediction4)
-
-    file_path = os.path.join(module_dir, 'model5.sav')
-    file5 = open(file_path, 'rb')   
-    m5 = pickle.load(file5)
-    file5.close()
-    prediction5 = m5.predict(userdata)
-    print(prediction5)
-    return (prediction5 + prediction2 + prediction3 + prediction4)/4
-    
-
 
 # SOS
 def SOS(request):
