--- templates/home/safety_check.html
@@ -1,69 +1,0 @@
-{% load static %}
-
-<!DOCTYPE html>
-<html>
-    <head>
-        <style>
-                *{
-                    box-sizing: border-box
-                }
-                body{
-                    margin: 0px;
-                    padding: 0px;
-                    
-                }
-            .place_details{
-                display:flex;
-                height: 100%;
-                border: 1px solid #e0e0e0;
-                margin: 10px 200px;
-                border-radius: 15px;
-                box-shadow: 0px 0px 5px 3px rgba(0,0,0,0.15);
-                padding: 10px;
-                background-color: #fff;
-            }
-            .place_detail{
-                display: inline;
-            }
-            .place_image{
-                width:40%;
-                padding: 0 15px;
-            }
-            .place_image img{
-                width:100%;
-                object-fit: cover;
-                object-position: center;
-                height: 200px;
-            }
-            .main {
-                background-color: aqua;
-            }
-        </style>
-    </head>
-    <body>
-    <div class="main">
-        <center><h1>Safety Check</h1></center>
-        {% for place in places%}
-        <div class="place_details">
-            <div class="place_image">
-                <!--<img src='https://unsplash.com/photos/KMn4VEeEPR8' class="hostel_image" alt="">-->
-                <img src="/media/{{ place.picture }}">
-            </div>
-
-            <div class="place_detail">
-                <h2>Name: {{ place.name }}</h2>
-                <br>
-                <button>
-                    <a href="{% url 'measure_safety' place.pk %}">Check Safety</a>
-                </button>
-            </div>
-
-        </div>
-        {% endfor %}
-    </div>
-    </body>
-    
-
-</html>
-
-
